feat: 전체/단일 게시글 조회 기능 추가

+++ src/main/java/org/example/buskmate/community/dto/crud/request/DeletePostRequest.java
@@ -1,9 +1,6 @@
 package org.example.buskmate.community.dto.crud.request;
 
-import org.example.buskmate.community.domain.CommunityPost;
-import org.example.buskmate.community.dto.crud.response.DeletePostResponse;
-
 public record DeletePostRequest(
-        Long id,
+        Integer id,
         String authorId
 ) { }
